Improve structure and documentation of SortBy connector
